+++ labelme/translate/it_IT.ts
@@ -137,6 +137,18 @@
         <translation>Clicca per aggiungere un punto</translation>
     </message>
     <message>
+        <source>Click &amp; drag to adjust profile width</source>
+        <translation>Click &amp; drag to adjust profile width</translation>
+    </message>
+    <message>
+        <source>Click &amp; drag to move width anchor</source>
+        <translation>Click &amp; drag to move width anchor</translation>
+    </message>
+    <message>
+        <source>Click &amp; drag to move visibility anchor</source>
+        <translation>Click &amp; drag to move visibility anchor</translation>
+    </message>
+    <message>
         <source>ALT + SHIFT + Click to delete point</source>
         <translation>ALT + MAIUSC + Clicca per eliminare il punto</translation>
     </message>